+++ embassy-time-queue-utils/src/queue_integrated.rs
@@ -47,6 +47,7 @@ unsafe impl Sync for Queue {}
 
 impl Queue {
     /// Creates a new timer queue.
+    #[inline]
     pub const fn new() -> Self {
         Self { head: Cell::new(None) }
     }
@@ -55,6 +56,7 @@ impl Queue {
     ///
     /// If this function returns `true`, the called should find the next expiration time and set
     /// a new alarm for that time.
+    #[inline]
     pub fn schedule_wake(&mut self, at: u64, waker: &Waker) -> bool {
         let item = unsafe {
             // Safety: the `&mut self`, along with the Safety note of the Queue, are sufficient to
@@ -68,6 +70,7 @@ impl Queue {
     ///
     /// If this function returns `true`, the called should find the next expiration time and set
     /// a new alarm for that time.
+    #[inline]
     pub fn schedule_wake_queue_item(&mut self, at: u64, item: &mut TimerQueueItem, waker: &Waker) -> bool {
         let item = unsafe { item.as_mut::<QueueItem>() };
         match item.waker.as_ref() {
@@ -106,6 +109,7 @@ impl Queue {
     ///
     /// The provided callback will be called for each expired task. Tasks that never expire
     /// will be removed, but the callback will not be called.
+    #[inline]
     pub fn next_expiration(&mut self, now: u64) -> u64 {
         let mut next_expiration = u64::MAX;