+++ embassy-time-queue-utils/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased - ReleaseDate
 
 - Both integrated and generic queue implementations are available for use, independent of their respective features.
+- Queue methods are now `#[inline]`
 
 ## 0.3.0 - 2025-08-26